refactor(aaa): rename tools -> versions
