+++ pallets/ajuna-awesome-avatars/src/types/avatar/versions/v1.rs
