+++ pallets/ajuna-awesome-avatars/src/types/avatar/versions/v2/avatar_utils.rs
