+++ pallets/ajuna-awesome-avatars/src/types/avatar/versions/v2/combinator/glimmer.rs
